Add Meals component for displaying meal items and headers
The new `Meals` component is added to display a list of meal items with headers. It includes `MealHeader` and `MealItem` sub-components for rendering the headers and individual items respectively. This component enhances the user interface by providing a visually appealing and organized way to display meals and their calorie information.