Update Storage module to use react-native-mmkv for data storage
The Storage module has been updated to use the react-native-mmkv library for data storage. This change enhances the performance and reliability of data storage operations by leveraging the efficient and secure features provided by react-native-mmkv. The module now includes functions for setting, getting, and removing items from storage, as well as a function to clear all stored data. This update improves the overall data management capabilities of the application.